PROVISORY: ci: cram: make exception for parameter change during LCM Agent restore
There is a bug when on the container start the RootfsIsMounted parameter
is set to false even though the container rootfs is actually mounted.
The bug is unstable. For this crma test in some cases it will show as
changed, in others it won't.

We add the parameter to the list of runtime parameters so we track its
changes (if any). In this way, the test won't fail if the parameter was
changed due to the bug and we didn't take it into account.

Implements: LCM-843
References: PPW-985

Signed-off-by: Gleb Kasimov <h.kasimov@inango-systems.com>