ci: cram: 065-upnp_0: update tests after introduction of tr181-upnpdiscovery
The current test fails while checking the amount of tr181-upnp processes that are running.
With the introduction of tr181-upnpdiscovery, we have now two processes matching the search pattern in stead of one.

So let's fix it by explicitly checking for tr181-upnp process and in addition let's add a test to check if tr181-upnpdiscovery is running as well.

References: PCF-1833

Signed-off-by: Florian Rouzaud <florian.rouzaud@softathome.com>