ci: testbed-cdrouter.py: add gitlab-branch and gitlab-job tags
When looking for regressions, it's handy to know on which branch/job the
test was run, so let's add support for branch/job tagging.

This commit introduces the `GitLabEnvironment` class, which extracts
GitLab CI environment variables related to branch names and job IDs.
It generates CDRouter-compatible tags that denote both the branch and
the CI job ID for the test execution. The priority for determining the
branch name is as follows:

1. CI_MERGE_REQUEST_SOURCE_BRANCH_NAME (for merge request pipelines)
2. CI_COMMIT_BRANCH (for regular branch pipelines)

The generated tags are formatted as:
- `gitlab-branch.{sanitized-branch}` for branch information
- `gitlab-job.{job_id}` for CI job identification

These tags are added to the existing tags generated by the
OpenWrtSystemInfo class and any user-defined tags. This addition aims
to simplify regression analysis by making it clear on which branch and
in which CI job a particular test execution occurred.

Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>
(cherry picked from commit addc70be8919a8b197e2ad45960fe57c56bf2308)
(cherry picked from commit 73a19798e9adf663504fb8fa28a9e7bd2455acc9)