profiles: prpl_core: feed-prplos: cleanup unnecessary packages
Some openwrt upstream packages were imported in feed-prplos for various reasons in the past:
- CVE fixing
- Import of newer versions because newer functionality was needed

This commit cleans up some packages from feed-prplos that are now by default the correct version in openwrt24.10.

Integrates:
* libarchive: remove package
The current version used in latest-24.10 is v3.7.9.
So it is safe to remove this import that was put in place to fix CVEs.

* libmbim: libqmi: remove imported packages
These newer versions of the libs were needed because we wanted to use a newer version of modemmanager in the past (v1.22.0).
Since we are now using openwrt-24.10, we already get newer versions of these libs that are compatible with modemmanager v1.22.0.
The modemmanager package cannot be removed, since we still need to update the dependency on the ppp package.

References: PPW-1098

Signed-off-by: Matthias Franck <matthias.franck@softathome.com>
(cherry picked from commit 224c2d3f60ef97ef9ccb52aa222009ff18ed93aa)
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>