profiles: prpl.yml: feed_prplos: reboot-service: bump to v1.0.7
The reboot_service module was previously tagged without the 'v' prefix

(e.g., 1.0.7 instead of v1.0.7), causing the Makefile to be generated
as if it were from the development branch. This resulted in an incorrect
Makefile and build failure.

Updated the tag to include the proper 'v' prefix and regenerated the
Makefile to ensure correct build behavior.

Signed-off-by: default avatarahmed sassi <ahmed.sassi_ext@softathome.com>
(cherry picked from commit 2603deb7f2247d279972a414171d119d39310c01)