Problem: tr181-captiveportal cram test are failing.
Cause: One of test case "get a DHCP lease from the router" output is not inline
with the actual output from the server and its only seen on our local
setup. On gitlab we are seeing that server is responding with the output
expected by the cram test case.

Solution: Reverting below commit as its not required and discripancy is
seen when cram test is run on local setup.

Revert "tr181-captiveportal:  Fix prpl cram test on 4.0.2"

This reverts commit cb53cc383fd2582e491cd36bb6463d3f9ab7a66b.

References: PPW-876
Signed-off-by: Manohar Singh <manohar.singh_ext@softathome.com>