ci: cram: lcm: pcm: add regexps to handle variable IP address changes of interfaces
A bug was introduced by running the test on the Arcadyan Mozart board.
```
@@ -93,6 +93,7 @@
   > EOF
   $ cthulhu_diff_params=$(diff -n ${TESTDIR}/cthulhu_before.dm ${TESTDIR}/cthulhu_after.dm | grep -o '^Cthulhu[^=]\+')
   $ for param in ${cthulhu_diff_params}; do grep -Fxq "${param}" ${TESTDIR}/runtime_params || echo "ERROR: runtime parameter mismatch - ${param}"; done
+  ERROR: runtime parameter mismatch - Cthulhu.Container.Instances.1.Interfaces.2.Addresses.1.Address
   $ timingila_diff_params=$(diff -n ${TESTDIR}/timingila_before.dm ${TESTDIR}/timingila_after.dm | grep -o '^SoftwareModules[^=]\+')
   $ for param in ${timingila_diff_params}; do grep -Fxq "${param}" ${TESTDIR}/runtime_params || echo "ERROR: runtime parameter mismatch - ${param}"; done
```
The board has a smaller number of network interfaces than the OSPv2
board, for example. This difference is reflected in the
Cthulhu.Container.Instances.{i}.Interfaces.{i} data model, specifically in
the number of instances of interfaces.

The test only handles the hardcoded value for the sixth interface
(lcm0), which was derived from testing on the OSPv2 board. However, the
Arcadyan Mozart board has fewer interfaces than the OSPv2 board, so the
interface index corresponding to lcm0 is shifted to the second position.

This commit introduces regular expressions to the list of runtime
parameters to track changes in parameters with variable instance
indexes. This commit also improves the user experience by outputting the
difference values before and after the fake firmware upgrade.

References: PPW-1136

Signed-off-by: Gleb Kasimov <h.kasimov@inango-systems.com>