PROVISORY: ci: cram : ospv2: 010-network.t: update ip link expected output
After the integration of the bSTAMLD feature, the default qdisc mode
for endpoint link interfaces changed from noqueue to noop.

As this mode is managed by the kernel/driver and has no impact on the
endpoint behavior when used, lets update the expected mode in the
script.

The ticket PPW-1282 will be used to analyze in more detail why the
mode changed.

References: FEAT-117, PPW-1282
Closes: PPM-3349
Signed-off-by: Houssem Dafdouf <houssem.dafdouf_ext@softathome.com>