ci: cram: wifi-usp.t: fix test order
Problem:
The wifi-usp.t script is impacting tests that verify the default wifi
interface configuration (network.t and wifi.t).

Cause:
The wifi-usp.t script, which recently started manipulating VAPs, is
being executed before network.t and wifi.t. Those tests implicitly
assume that VAPs have never been enabled, causing the precondition to
be broken.

Solution:
Run the wifi-usp.t script at the appropriate time after network.t and
wifi.t.

Note:
The wifi-usp.t script was being executed twice because it existed in
the generic/ folder and also a symlink is pointing to it in the DUT
folder.
To resolve this, and by the way the initial purpose of this commit,
let's move it under wifi/ folder and update symlinks accordingly.

Closes: PPW-1259
References: PPW-1208
Signed-off-by: Houssem Dafdouf <houssem.dafdouf_ext@softathome.com>