PROVISORY: ci: freedom: fix WPS paring timeout issue
Problem:

During the EM R1 certificaiton mode the WPS pairing ends always with a
time out.

Cause:

The issue is caused by an unexpected pwhm behavior when setting
regulatory domain:

- After a firstboot all access points are disabled
- Endpoints are enabled, so radios are in Dormant state
- When changing the regulatory domain the radio becomes Down, and this
is an unexpected as radio should be kept Dormant
- Later when triggering a WPS paring, the push button RPC visibly
starts the pairing process, but nothing is happening on radio side (not
 beaconing). As a consequence the pairing will time out.

Solution:

Toggle the endpoint status after updating the regulatory domain so
radio can be reactivated.

References: PPW-1326
Signed-off-by: Houssem Dafdouf <houssem.dafdouf_ext@softathome.com>