ci: cram: update 006-amx-monitoring-change-subject.t testcase (PPW-1228)
Problem - Failure action as set in FailAction is taken when
repeated process kill exceeds MaxFailNum parameter.

Process respawn parameters were tested in previous
005-amx-respawn-reset.t testcase which upon failure increases the
CurrentTestInterval. ProcessMonitor detects the Health of the
process based on this CurrentTestInterval.

This CurrentTestInterval after failure was around 120 seconds for
tr181-qos process. Increased time for tr181-qos process observed
would be due to some small varying additional time taken by it to
come up at the expected monitoring interval (possibly due to
additional tasks handled by tr181-qos process compared to other process)
for example tr181-pcp process had CurrentTestInterval as 60 seconds as
per 006-amx-monitoring-change-subject.t.err file.

Due to this reason the issue reported in PPW-1228 is observed. Here
NumProcessRespawn for tr181-qos was expected as 1 but found 0 because
FailAction was invoked while Health was still in Bad state, RESTART
FailAction for tr181-qos was taken which respawns the process and
clears NumProcessRespawn and MaxNumFailed attributes.

Solution - Instead of waiting for the increased CurrentTestInterval
for the Health to trun into Good state and proceed with further
testcases, TestInterval parameter is reduced to 10 seconds from 60
seconds. This improves updating the Health status of the process.
Additionally MaxFailNum is increased to higher value at the start of
the tests to avoid unexpected FailActions and reverted at the end of
testcase. Also adding update to log CurrentTestInterval and Health
for all amx-processmonitoring cram testcases.

Fixes: PPW-1228
Signed-off-by: Hari Sriram <hari.s@capgemini.com>